+++ src/test/data/object/deepClone.js
@@ -1,8 +1,10 @@
 
+import text from '../../main';
 import updateData from './../../../data/object/updateData'
 import deepClone from './../../../data/object/deepClone'
+import mergeData from '../../../data/object/mergeData';
 
-(function() {
+text(function() {
   // 拷贝相关
   let targetdata = {
     name: 'target',
@@ -38,8 +40,8 @@ import deepClone from './../../../data/object/deepClone'
   if (targetdata.name != 'origin' || targetdata.data.list[0].name != '11' || targetdata.data.list[1].id != 3) {
     console.error('UpdateData未成功')
   }
-})();
-(function() {
+});
+text(function() {
   // 拷贝相关
   let targetdata = {
     name: 'target',
@@ -86,24 +88,72 @@ import deepClone from './../../../data/object/deepClone'
   if (targetdata.data.list !== origindata.data.list) {
     console.error('UpdateData深度限制未成功')
   }
-})();
+});
 
-(function() {
+text(function() {
   // 拷贝相关
-  try {
-    let obj = {
-      a: 1,
-      b: [1, 2, 3],
-      c: {
-        h: 20
-      },
-      d: () => {}
-    }
-    obj.b.push(obj.c)
-    obj.c.j = obj.b
-    deepClone(obj, true)
-    deepClone(obj, {})
-  } catch (e) {
-    console.error('深拷贝的循环引用报错', e)
+  let obj = {
+    a: 1,
+    b: [1, 2, 3],
+    c: {
+      h: 20
+    },
+    d: () => {}
   }
-})();
+  obj.b.push(obj.c)
+  obj.c.j = obj.b
+  deepClone(obj, true)
+  deepClone(obj, {})
+}, '深拷贝的循环引用报错');
+
+text(function() {
+  // 拷贝相关
+  let data = {
+    name: 'id',
+    prop: 'id',
+    mod: {
+      list: {
+        width: 100,
+        option: {
+          min: 0,
+          list: [1, 2, 3],
+          objectList: [
+            {
+              name: 1
+            },
+            {
+              name: 2
+            }
+          ]
+        }
+      },
+      edit: {
+        option: {
+          name: 'name'
+        }
+      }
+    }
+  }
+  let newdata = updateData(data, {
+    name: 'ID',
+    originprop: 'userId',
+    mod: {
+      list: {
+        local: true,
+        option: {
+          min: 1,
+          max: 100,
+          objectList: [1, {
+            name: '111'
+          }]
+        }
+      },
+      build: {
+        option: {
+          name: 'build'
+        }
+      }
+    }
+  })
+  console.log(newdata)
+});