+++ src/test/main.js
@@ -1,0 +1,13 @@
+
+function text(fn, log) {
+  try {
+    fn()
+  } catch (e) {
+    console.error(e)
+    if (log) {
+      console.error(log)
+    }
+  }
+}
+
+export default text