+++ history.md
@@ -2,6 +2,7 @@
 ### 2.0.19ing
 - 测试结构更新
 - formatTree逻辑优化，删除formatTreeNext,测试用例添加
+- updateList中option.format不返回需要push的数据，因为此时需要的对象都是复杂格式，内存指针形式，因此format仅对对象数据做格式化，返回值为是否添加到数组中
 
 ### 2.0.18
 - current注释